refactor skill actions to use i64 for better precision and update UI to display actions for skill levels

+++ src/game.rs
@@ -144,7 +144,7 @@ impl GameState {
                             self.inventory.add_item(*item, *amount);
                         }
                         Effect::IncrementActionsForSkill { skill_type, amount } => {
-                            self.skills.get_skill_by_type_mut(skill_type).increment_actions(*amount as u32);
+                            self.skills.get_skill_by_type_mut(skill_type).increment_actions(*amount);
                         }
                     }
